♻️ Adapt changes

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('user')->unique();
             $table->string('name');
             $table->string('password');
-            $table->decimal('salary',9,3);
-            $table->decimal('profit_margin',9,3);
+            $table->decimal('salary',9,2);
+            $table->integer('profit_margin');
             $table->unsignedBigInteger("role_id");
             $table->boolean("is_active")->default(true);
             $table->rememberToken();
